+++ types/params.go
@@ -2,6 +2,6 @@ package types
 
 const (
 	FeeFraction float64 = 0.005 // 0.5%
-	MinimumFee  uint64  = 1e9   // 1 PAC
-	MaximumFee  uint64  = 5e9   // 5 PAC
+	MinimumFee  float64 = 1e9   // 1 PAC
+	MaximumFee  float64 = 5e9   // 5 PAC
 )
